combine and conquer (vercel cmon man)
