--- api/clubs/leave.js
@@ -1,17 +1,0 @@
-import neo4j from "neo4j-driver";
-
-let driver;
-try {
-    driver = neo4j.driver(
-        process.env.NEO4J_URI,
-        neo4j.auth.basic(process.env.NEO4J_USERNAME, process.env.NEO4J_PASSWORD)
-    );
-} catch (error) {
-    console.error('Failed to create driver:', error);
-}
-
-export default async function handler(req, res) {
-    if (req.method !== 'POST') {
-        return res.status(405).json({ error: 'Method not allowed' });
-    }
-}
